first step into picking weapon and reset bug
Almost done everything for a player to switch weapon with the one on the
cell but also discovered a bug with the reset : if you start a new game
when red (first) is playing, it will completly break the turn system for
the third turn of the new game. To be continued.