Responsive work
It was about time but the design is now 90% responsive, I just nedd to
do a little bit more work on the players pictures (in md  size) and the
buttons around the main title (in xs size)
I made a test : the grid is fully responsive at the sacrifice that the
cells are too, so they are not always square. I need to test it a bit to
see the pro and the cons of this solution.